+++ client/src/components/AuthHeader/AuthHeader.tsx
@@ -1,7 +1,7 @@
 import Button from '@material-ui/core/Button';
 import Box from '@material-ui/core/Box';
 import { Link } from 'react-router-dom';
-import useStyles from './useStyles';
+import useStyles from './authHeaderStyles';
 import { Typography } from '@material-ui/core';
 
 interface Props {
