+++ client/src/components/ChatSideBanner/ChatSideBanner.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, useState } from 'react';
 import Box from '@material-ui/core/Box';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import useStyles from './useStyles';
+import useStyles from './chatSideBannerStyles';
 import { User } from '../../interface/User';
 import AvatarDisplay from '../AvatarDisplay/AvatarDisplay';
 import Search from '../Search/Search';
